+++ src/error/prelude.rs
@@ -1,0 +1,5 @@
+pub use super::{Error, Result};
+pub use std::io::Error as IoError;
+pub use std::io::ErrorKind as IoErrorKind;
+pub use serde_json::error::Error as JsonError;
+pub use super::service_error::ServiceError;